backend/requirements.txt

django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
reportlab
dj-database-url
psycopg2-binary
gunicorn
